GPU/requirements.txt
pika==1.3.2
boto3==1.34.68
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.8.1
tqdm==4.66.5
ultralytics==8.3.78
onnx==1.17.0
onnxruntime-gpu==1.19.0